adding all the missing FadeAnimation on text and contianer and columns and fixing the const missing in the Text widgets

+++ lib/Animation/FadeAnimation.dart
@@ -4,7 +4,7 @@ class FadeAnimation extends StatelessWidget {
   final Widget child;
   final Duration duration;
 
-  FadeAnimation(int i, 
+  FadeAnimation(double i,
       {required this.child, this.duration = const Duration(milliseconds: 500)});
 
   @override
